+++ Packages/Gr8tGames.Audio/Runtime/MinMaxAudioSettingAttribute.cs
@@ -1,13 +1,13 @@
-﻿namespace GD.MinMaxSlider
+﻿namespace Gr8tGames.Audio
 {
     using UnityEngine;
 
-    public class MinMaxSliderAttribute : PropertyAttribute{
+    public class MinMaxAudioSettingAttribute : PropertyAttribute{
 
         public float min;
         public float max;
 
-        public MinMaxSliderAttribute(float min, float max)
+        public MinMaxAudioSettingAttribute(float min, float max)
         {
             this.min = min;
             this.max = max;